AI Release Testing Agent/requirements.txt
# Core web framework
flask>=3.0.0

# OpenAI SDK for AI-powered test generation (v1.x+ API)
openai>=1.0.0

# Google Gemini SDK for AI-powered test generation
google-genai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests (utility)
requests>=2.31.0
